明示的な型変換を行ってから計算

+++ ITP_1_4_A.cpp
@@ -8,7 +8,7 @@ int main(){
     cin >> a >> b;
     r = a/b;
     d = a%b;
-    f = a/b;
+    f = (double)a/(double)b;
     cout << r << " " << d << " " << f << endl;
     return 0;
 }
